Inline and decouple test and data loading code from mbi_tests.py into Locustfiles that used it; remove mbi_tests.py
